+++ models.py
@@ -1,4 +1,6 @@
+import asyncio
 import datetime
+from typing import Dict
 from typing import Optional
 
 import beanie
@@ -91,3 +93,68 @@ class HealthCareData(beanie.Document):
                 ]
             ),
         ]
+
+
+class Borg:
+    """Borg inspired singleton class."""
+
+    _shared_state = {}
+
+    def __init__(self):
+        self.__dict__ = self._shared_state
+
+
+class PsqlClient(Borg):
+    """Creates a singleton postgresql client instance.
+
+    This creates all postgres tables for a database only once.
+    To be used during startup event of a asgi server or
+    with dependency injection.
+    """
+
+    __instance: Optional[Borg] = None
+
+    def __new__(cls, model: orm.ModelRegistry, *args, **kwargs):
+        if not cls.__instance:
+            cls.__instance = super().__new__(cls)
+        return cls.__instance
+
+    def __init__(self, model: orm.ModelRegistry):
+        super().__init__()
+        self.model = model
+
+    async def create_async_model(self):
+
+        url = self.model._get_database_url()
+        tasks = [
+            # See: https://github.com/encode/databases/issues/96#issuecomment-619425993
+            asyncio.create_task(self.model.database.connect()),
+            # Probably a fix.
+            # See: https://github.com/encode/orm/issues/114#issuecomment-974735799
+            asyncio.create_task(self.model._create_all(url)),
+        ]
+        await asyncio.gather(*tasks)
+
+    async def drop_async_model(self):
+        url = self.model._get_database_url()
+        await self.model._drop_all(url)
+        await self.model.database.disconnect()
+
+
+async def start_psql_client(model_registry: orm.ModelRegistry):
+    await PsqlClient(model_registry).create_async_model()
+
+
+async def drop_psql_client(model_registry: orm.ModelRegistry):
+    await PsqlClient(model_registry).drop_async_model()
+
+
+def get_postgres_collections(
+    models_registry: orm.ModelRegistry,
+) -> Dict[str, orm.Model]:
+    """Returns a dictionary of collection name with orm object."""
+    tables: Dict[str, orm.ModelRegistry] = models_registry.models
+    return {
+        orm_registry_model.tablename: orm_registry_model
+        for orm_registry_model in tables.values()
+    }
